Added Form Validation
You can click on all Coins except for 1 cent to change it's value. Form
validation is also performed.
The total form validation is performed. However, due to performance
issues, I've limited the max total to be 1000. Since the app now is pure
front-end angularjs, the calculation is done by the brower's javascript
engine, which is limited.
Calculation can be moved to backend so performance will be boosted.